Replaces animated header with shared navbar. Adds navbar to default layout and removes LayoutWithNavigation.
